+++ accounting/src/sdk/purchaseorders.ts
@@ -19,7 +19,7 @@ export enum DownloadAttachmentAcceptEnum {
 
 export enum DownloadPurchaseOrderPdfAcceptEnum {
     applicationJson = "application/json",
-    applicationOctetStream = "application/octet-stream",
+    applicationPdf = "application/pdf",
 }
 
 export class PurchaseOrders {
@@ -336,7 +336,7 @@ export class PurchaseOrders {
         if (acceptHeaderOverride !== undefined) {
             headers["Accept"] = acceptHeaderOverride.toString();
         } else {
-            headers["Accept"] = "application/json;q=1, application/octet-stream;q=0";
+            headers["Accept"] = "application/json;q=1, application/pdf;q=0";
         }
 
         headers["user-agent"] = this.sdkConfiguration.userAgent;
@@ -380,7 +380,7 @@ export class PurchaseOrders {
         const decodedRes = new TextDecoder().decode(httpRes?.data);
         switch (true) {
             case httpRes?.status == 200:
-                if (utils.matchContentType(responseContentType, `application/octet-stream`)) {
+                if (utils.matchContentType(responseContentType, `application/pdf`)) {
                     res.data = httpRes?.data;
                 } else {
                     throw new errors.SDKError(